fix(phase8): resolve suboptimal-region + MAPQ parity tail (SE 5000/5000, PE 9999/10000)
Diagnosed against a bit-identical instrumented oracle (Homebrew sse2neon arm64
build reproduced in scratch, byte-identical to the installed binary except PG),
dumping SMEMs, pre/post-mem_chain_flt chains, post-mark regions and the
mem_approx_mapq_se inputs. Two root causes:

1. Chain order + unstable sort in mem_chain_flt (340 SE + 622 PE records, all
   suboptimal regions at a different locus). bwa-mem2 flattens a position-keyed
   kbtree (chains ascending by pos) then applies an *unstable* ks_introsort by
   weight-descending; for equal-weight query-overlapping chains the survivor
   depends on input order and the exact introsort permutation. We built chains
   in occurrence order and used a stable sort. Fix: build_chains sorts by pos;
   mem_chain_flt uses a faithful ks_introsort port (median-of-3 quicksort +
   combsort depth fallback + final insertion sort). The primary (max weight) is
   always selected identically.

2. mapQ_coef_fac is an int in bwa-mem2 ((int)log(50) = 3, not 3.912). We stored
   the float ln(50), inflating MAPQ on borderline reads (23 SE + 6 PE, MAPQ-only,
   regions otherwise identical). Fix: truncate to the integer value.

Verified no regression, -t8 == -t1 byte-identical (excluding PG). Remaining: 1
PE record, cosmetic XS only (MAPQ=60 intact), a distinct SW-extension case on a
repetitive locus where the suboptimal chain's seed composition differs.

+++ crates/bwa-core/src/opt.rs
@@ -55,7 +55,9 @@ pub struct MemOpt {
     pub mask_level_redun: f32,
     /// MAPQ coefficient length.
     pub mapq_coef_len: f64,
-    /// MAPQ coefficient factor (`ln(mapq_coef_len)`).
+    /// MAPQ coefficient factor. bwa-mem2 declares this `int` and sets it to `(int)log(mapq_coef_len)`,
+    /// so the fractional part is truncated (`log(50) = 3.912` becomes `3`). We keep it as an `f64`
+    /// holding that already-truncated integer value; the truncation matters for borderline MAPQs.
     pub mapq_coef_fac: f64,
     /// Max insert size.
     pub max_ins: i32,
@@ -105,7 +107,8 @@ impl Default for MemOpt {
             xa_drop_ratio: 0.80,
             mask_level_redun: 0.95,
             mapq_coef_len: 50.0,
-            mapq_coef_fac: 50.0f64.ln(),
+            // (int)log(50) = 3, matching bwa-mem2's integer `mapQ_coef_fac`.
+            mapq_coef_fac: (50.0f64.ln() as i32) as f64,
             max_ins: 10000,
             max_matesw: 50,
             max_mem_intv: 20,
